requirements.txt
crawl4ai>=0.3.0
celery[redis]>=5.3.0
psycopg2-binary>=2.9.0
pandas>=2.1.0
openpyxl>=3.1.0
playwright>=1.40.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0